Create classes for the entities we need from what me and Doddi talked about at 17.2.2016
So I created a template for the Airport.class, Costumer.class, Flights.class, Plane.class. 

I created them with default constructors and fully functional constructors for adding new entities. And also created a template for new methods we need to implement to get the search operations running. 

So now it is relatively easy to create airports, costumers and planes. But I have not finished the getters and setters for the Flights.class because i think it needs more thought. 

So check this out and try to test it out using the test section at the bottom of each class.